[SYCL] Workaround a circular reference
detail/helpers.hpp was needed by spirv_ops.cpp for the inline definition
of detail::memcpy, but including such caused a circular ordering issue.

There's no need to forward declare an inline function that's immediately
needed, so hoist it into its own header that can easily be included in
both users.
